# Speed up hiding computations by exiting early if the sphere is far enough to the right or left of the segment