Before:
```
BM_VisibleSegmentsOrbit/10                        3272 ns       3285 ns     213674 average visible segments: 1.000000
BM_VisibleSegmentsOrbit/100                      31592 ns      31289 ns      22436 average visible segments: 0.980000
BM_VisibleSegmentsOrbit/1000                    312704 ns     310417 ns       2362 average visible segments: 0.966000
BM_VisibleSegmentsRandomEverywhere/1000         136836 ns     137666 ns       4986 average visible segments: 1.073000
BM_VisibleSegmentsRandomNoIntersection/1000     106438 ns     102207 ns       7479 average visible segments: 1.000000
```
After:
```
BM_VisibleSegmentsOrbit/10                        1835 ns       1836 ns     407923 average visible segments: 1.000000
BM_VisibleSegmentsOrbit/100                      11361 ns      11125 ns      56089 average visible segments: 0.980000
BM_VisibleSegmentsOrbit/1000                    109154 ns     109517 ns       6410 average visible segments: 0.966000
BM_VisibleSegmentsRandomEverywhere/1000         119925 ns     119594 ns       5609 average visible segments: 1.073000
BM_VisibleSegmentsRandomNoIntersection/1000      89517 ns      90395 ns       8974 average visible segments: 1.000000
```